Remove Xamarin support
