+++ LibsAndSamples.sln
@@ -155,8 +155,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "docs", "docs", "{38B77C31-B
 		ThirdPartyNotice.txt = ThirdPartyNotice.txt
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Identity.Test.UIAutomation.Infrastructure", "tests\Microsoft.Identity.Test.Core.UIAutomation\Microsoft.Identity.Test.UIAutomation.Infrastructure.csproj", "{3DC6EC76-D350-4D43-B206-A4CEC8AA36D4}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Managed Identity apps", "Managed Identity apps", "{BCAEE9AE-8D3E-4C77-A2E4-134E1552D5F8}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ManagedIdentityAppVM", "tests\devapps\Managed Identity apps\ManagedIdentityAppVM\ManagedIdentityAppVM.csproj", "{46DF28CE-D680-4E9A-9871-95A1C42DD15D}"
@@ -1927,7 +1925,6 @@ Global
 		{6839F934-45F0-4026-8AF3-C3AEFB7D48A9} = {1A37FD75-94E9-4D6F-953A-0DABBD7B49E9}
 		{69CE42A2-44DD-4740-856B-607C3BB5A92E} = {9B0B5396-4D95-4C15-82ED-DC22B5A3123F}
 		{259B5734-1D43-43F7-8107-32F602A99019} = {EA983957-DC2F-4D9E-A2E3-79ADEB5CEB26}
-		{3DC6EC76-D350-4D43-B206-A4CEC8AA36D4} = {9B0B5396-4D95-4C15-82ED-DC22B5A3123F}
 		{BCAEE9AE-8D3E-4C77-A2E4-134E1552D5F8} = {34BE693E-3496-45A4-B1D2-D3A0E068EEDB}
 		{46DF28CE-D680-4E9A-9871-95A1C42DD15D} = {BCAEE9AE-8D3E-4C77-A2E4-134E1552D5F8}
 		{128D16A2-1B07-4600-A898-71B328375440} = {5FAAD966-36B8-4C19-A5FA-5410DD53063D}
