+++ tests/Microsoft.Identity.Test.Integration.netfx/Infrastructure/SeleniumExtensions.cs
@@ -9,7 +9,6 @@ using OpenQA.Selenium.Chrome;
 using OpenQA.Selenium.Support.UI;
 using System.Diagnostics;
 using Microsoft.Identity.Test.LabInfrastructure;
-using Microsoft.Identity.Test.UIAutomation.Infrastructure;
 using System.Linq;
 using Microsoft.Identity.Client;
 using System.Threading;
